backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.14.3

# Redis
redis[hiredis]==5.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP client


# AI
google-generativeai==0.4.1

# Firebase
firebase-admin==6.4.0

# Supabase
supabase>=2.3.4

# Middleware
python-dotenv==1.0.1

# Scheduling
apscheduler==3.10.4

# CSV
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

email-validator>=2.0.0